Change U2P_PlayerSettings to have a non-static initialization method instead of a static factory method. Initialize it at the beginning of U2P_PlayerState.
The idea is that U2P_PlayerState will manage the only instance of U2P_PlayerSettings. The settings will stay in a separate object because CVars, and therefore U2P_PlayerSettings, have a different lifecycle (they aren't saved, and are reinitialized on loading a save) than U2P_PlayerState (which is saved).